requirements.txt
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.0
gym==0.21.0
gym-notices==0.0.8
importlib-metadata==4.13.0
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
MouseInfo==0.1.3
mss==7.0.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pipreq==0.4
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.10
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.4
pytz==2022.7.1
six==1.16.0
stable-baselines3==1.7.0
torch==1.13.1
typing_extensions==4.5.0
zipp==3.15.0
pynput==1.7.6